Bumped rack gem to get rid of CVE-2018-16471 vulnerability:
Vulnerable versions: >= 2.0.0, < 2.0.6
Patched version: 2.0.6
There is a possible XSS vulnerability in Rack before 2.0.6 and 1.6.11. Carefully crafted requests can impact the data returned by the scheme method on Rack::Request. Applications that expect the scheme to be limited to 'http' or 'https' and do not escape the return value could be vulnerable to an XSS attack. Note that applications using the normal escaping mechanisms provided by Rails may not impacted, but applications that bypass the escaping mechanisms, or do not use them may be vulnerable.

Added `:frozen_string_literal` pragma to Gemfile, thus making Rubocop completely happy.